@@ -24,16 +24,12 @@ void func(void)
 if __name__ == '__main__':
     parser = c_parser.CParser()
     ast = parser.parse(text)
+    dump_filename = 'ast.pickle'
 
-    # Since AST nodes use __slots__ for faster attribute access and
-    # space saving, it needs Pickle's protocol version >= 2.
-    # The default version is 3 for python 3.x and 1 for python 2.7.
-    # You can always select the highest available protocol with the -1 argument.
-
-    with open('ast', 'wb') as f:
-        pickle.dump(ast, f, protocol=-1)
+    with open(dump_filename, 'wb') as f:
+        pickle.dump(ast, f, protocol=pickle.HIGHEST_PROTOCOL)
 
     # Deserialize.
-    with open('ast', 'rb') as f:
+    with open(dump_filename, 'rb') as f:
         ast = pickle.load(f)
         ast.show()
