@@ -111,7 +111,9 @@ class CLexer(object):
     keywords_new = (
         '_BOOL', '_COMPLEX',
         '_NORETURN', '_THREAD_LOCAL', '_STATIC_ASSERT',
-        '_ATOMIC', '_ALIGNOF', '_ALIGNAS', '_GENERIC',
+        '_ATOMIC', '_ALIGNOF', '_ALIGNAS',
+        '_Generic',
+        '_PRAGMA',
         )
 
     keyword_map = {}
