@@ -571,14 +571,28 @@ class CParser(PLYParser):
         self._parse_error('Directives not supported yet',
                           self._token_coord(p, 1))
 
+    # This encompasses two types of C99-compatible pragmas:
+    # - The #pragma directive:
+    #       # pragma character_sequence
+    # - The _Pragma unary operator:
+    #       _Pragma ( " string_literal " )
     def p_pppragma_directive(self, p):
         """ pppragma_directive      : PPPRAGMA
                                     | PPPRAGMA PPPRAGMASTR
+                                    | _PRAGMA LPAREN unified_string_literal RPAREN
         """
-        if len(p) == 3:
+        if len(p) == 5:
+            p[0] = c_ast.Pragma(p[3], self._token_coord(p, 2))
+        elif len(p) == 3:
             p[0] = c_ast.Pragma(p[2], self._token_coord(p, 2))
         else:
             p[0] = c_ast.Pragma("", self._token_coord(p, 1))
+
+    def p_pppragma_directive_list(self, p):
+        """ pppragma_directive_list : pppragma_directive
+                                    | pppragma_directive_list pppragma_directive
+        """
+        p[0] = [p[1]] if len(p) == 2 else p[1] + [p[2]]
 
     # In function definitions, the declarator can be followed by
     # a declaration list, for old "K&R style" function definitios.
@@ -671,12 +685,12 @@ class CParser(PLYParser):
     #       sum += 1;
     #   }
     def p_pragmacomp_or_statement(self, p):
-        """ pragmacomp_or_statement     : pppragma_directive statement
+        """ pragmacomp_or_statement     : pppragma_directive_list statement
                                         | statement
         """
-        if isinstance(p[1], c_ast.Pragma) and len(p) == 3:
+        if len(p) == 3:
             p[0] = c_ast.Compound(
-                block_items=[p[1], p[2]],
+                block_items=p[1]+[p[2]],
                 coord=self._token_coord(p, 1))
         else:
             p[0] = p[1]
