@@ -1736,6 +1736,7 @@ class TestCParser_fundamentals(TestCParser_base):
             struct s {
             #pragma baz
             } s;
+            _Pragma("other \"string\"")
             '''
         s1_ast = self.parse(s1)
         self.assertIsInstance(s1_ast.ext[0], Pragma)
@@ -1758,6 +1759,10 @@ class TestCParser_fundamentals(TestCParser_base):
         self.assertEqual(s1_ast.ext[2].type.type.decls[0].string, 'baz')
         self.assertEqual(s1_ast.ext[2].type.type.decls[0].coord.line, 13)
 
+        self.assertIsInstance(s1_ast.ext[3], Pragma)
+        self.assertEqual(s1_ast.ext[3].string.value, r'"other \"string\""')
+        self.assertEqual(s1_ast.ext[3].coord.line, 15)
+
     def test_pragmacomp_or_statement(self):
         s1 = r'''
             void main() {
@@ -1776,6 +1781,7 @@ class TestCParser_fundamentals(TestCParser_base):
 
                 if (sum > 10)
                     #pragma bar
+                    #pragma baz
                     sum = 10;
 
                 switch (sum)
@@ -1800,7 +1806,8 @@ class TestCParser_fundamentals(TestCParser_base):
         self.assertIsInstance(s1_ast.ext[0].body.block_items[4], If)
         self.assertIsInstance(s1_ast.ext[0].body.block_items[4].iftrue, Compound)
         self.assertIsInstance(s1_ast.ext[0].body.block_items[4].iftrue.block_items[0], Pragma)
-        self.assertIsInstance(s1_ast.ext[0].body.block_items[4].iftrue.block_items[1], Assignment)
+        self.assertIsInstance(s1_ast.ext[0].body.block_items[4].iftrue.block_items[1], Pragma)
+        self.assertIsInstance(s1_ast.ext[0].body.block_items[4].iftrue.block_items[2], Assignment)
         self.assertIsInstance(s1_ast.ext[0].body.block_items[5], Switch)
         self.assertIsInstance(s1_ast.ext[0].body.block_items[5].stmt.stmts[0], Compound)
         self.assertIsInstance(s1_ast.ext[0].body.block_items[5].stmt.stmts[0].block_items[0],
